[282. Expression Add Operators][Compile Error]committed by LeetCode Extension

+++ 282-Expression-Add-Operators/solution.java
@@ -7,7 +7,7 @@ public class Solution {
         return res;
     }
     
-    private void helper(String str, char[] chs, int index, String path, List<String> res, int val, int target, int mul) {
+    private void helper(String str, char[] chs, int index, String path, List<String> res, long val, int target, long mul) {
         if (index == chs.length) {
             if (val == target) {
                 res.add(path);
@@ -25,13 +25,13 @@ public class Solution {
         } else {
             for (int i = index;i < chs.length;i++) {
                 String s = str.substring(index, i + 1);
-                int num = Integer.parseInt(s);
+                long num = Long.parseInt(s);
                 if (index == 0) {
                     helper(str, chs, i + 1, s, res, num, target, num);
                 } else {
                     helper(str, chs, i + 1, path + "+" + s, res, val + num, target, num);
                     helper(str, chs, i + 1, path + "-" + s, res, val - num, target, - num);
-                    int m = mul * num;
+                    long m = mul * num;
                     helper(str, chs, i + 1, path + "*" + s, res, val - mul + m, target, m);
                 }
             }
